# Conflicts:
#	Session.xcodeproj/project.pbxproj
#	Session/Conversations/ConversationViewModel.swift
#	Session/Conversations/Settings/ThreadSettingsViewModel.swift
#	Session/Home/HomeViewModel.swift
#	Session/Onboarding/LandingScreen.swift
#	Session/Settings/SettingsViewModel.swift
#	Session/Utilities/UIContextualAction+Utilities.swift
#	SessionMessagingKit/LibSession/Config Handling/LibSession+Contacts.swift
#	SessionMessagingKit/LibSession/Config Handling/LibSession+UserProfile.swift
#	SessionMessagingKit/Sending & Receiving/MessageSender+Convenience.swift
#	SessionMessagingKit/SessionPro/SessionProManager.swift
#	SessionMessagingKit/Types/ConversationInfoViewModel.swift
#	SessionUIKit/Components/SwiftUI/DonationCTAModal.swift
#	SessionUIKit/Components/SwiftUI/Modal+SwiftUI.swift
#	SessionUIKit/Components/SwiftUI/ProCTAModal.swift
#	SessionUIKit/Screens/SessionListScreen/ListItemViews/SessionListScreen+ListItemProfilePicture.swift
#	SessionUIKit/Screens/SessionListScreen/SessionListScreen.swift
#	SessionUIKit/Style Guide/Themes/SwiftUI+Theme.swift
#	SessionUtilitiesKit/Database/Models/KeyValueStore.swift

+++ SessionMessagingKit/LibSession/Config Handling/LibSession+UserGroups.swift
@@ -263,19 +263,19 @@ internal extension LibSessionCacheType {
                 {
                     // Add in any new members and remove any removed members
                     try updatedMembers.forEach { try $0.upsert(db) }
-                    try existingMembers
-                        .filter { !updatedMembers.contains($0) }
-                        .forEach { member in
-                            try GroupMember
-                                .filter(
-                                    GroupMember.Columns.groupId == group.id &&
-                                    GroupMember.Columns.profileId == member.profileId && (
-                                        GroupMember.Columns.role == GroupMember.Role.standard ||
-                                        GroupMember.Columns.role == GroupMember.Role.zombie
-                                    )
+                    try existingMembers.forEach { member in
+                        guard !updatedMembers.contains(member) else { return }
+                        
+                        try GroupMember
+                            .filter(
+                                GroupMember.Columns.groupId == group.id &&
+                                GroupMember.Columns.profileId == member.profileId && (
+                                    GroupMember.Columns.role == GroupMember.Role.standard ||
+                                    GroupMember.Columns.role == GroupMember.Role.zombie
                                 )
-                                .deleteAll(db)
-                        }
+                            )
+                            .deleteAll(db)
+                    }
                 }
 
                 if
@@ -1071,8 +1071,9 @@ public extension LibSession.Cache {
     }
     
     func groupInfo(for groupIds: Set<String>) -> [LibSession.GroupInfo?] {
-        guard case .userGroups(let conf) = config(for: .userGroups, sessionId: userSessionId) else { return [] }
-        
+        guard case .userGroups(let conf) = config(for: .userGroups, sessionId: userSessionId) else {
+            return []
+        }
         
         return groupIds.map { groupId -> LibSession.GroupInfo? in
             var group: ugroups_group_info = ugroups_group_info()
